+++ src/main/java/cart/application/MemberCouponService.java
@@ -8,6 +8,7 @@ import cart.domain.member.Member;
 import cart.dto.coupon.MemberCouponResponse;
 import cart.exception.BadRequestException;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 import java.util.stream.Collectors;
@@ -15,6 +16,7 @@ import java.util.stream.Collectors;
 import static cart.exception.ErrorCode.ALREADY_ISSUED_COUPON;
 
 @Service
+@Transactional(readOnly = true)
 public class MemberCouponService {
 
     private final CouponRepository couponRepository;
