+++ src/main/java/cart/domain/product/Item.java
@@ -1,4 +1,4 @@
-package cart.domain;
+package cart.domain.product;
 
 public class Item {
