+++ src/main/java/cart/dto/product/ItemRequest.java
@@ -1,4 +1,4 @@
-package cart.dto;
+package cart.dto.product;
 
 public class ItemRequest {
